Refactor code that loads attribute data by scope
- Having \Magento\Eav\Model\ResourceModel\Attribute\ScopeDataLoader::getAllScopeDataByAttribute
  and \Magento\Eav\Model\ResourceModel\ReadHandler::execute as separate
  methods is fair cleaner than the combined approach that was initially
  being used.
