+++ app/code/Magento/Catalog/Api/CategoryScopeManagementInterface.php
@@ -12,22 +12,9 @@ namespace Magento\Catalog\Api;
 interface CategoryScopeManagementInterface
 {
     /**
-     * Get category data for all scopes in array of following format
+     * Get category data for all scopes
      *
-     * ```php
-     * [
-     *     attribute_code_1 => [
-     *         store ID X => value of attribute_code_1 for store ID X,
-     *         store ID Y => value of attribute_code_1 for store ID Y,
-     *         ...
-     *     ],
-     *     attribute_code_2 => [
-     *         store ID X => value of attribute_code_2 for store ID X,
-     *         store ID Y => value of attribute_code_2 for store ID Y,
-     *         ...
-     *     ],
-     * ]
-     * ```
+     * @see \Magento\Eav\Model\ResourceModel\Attribute\DataLoader::getAllScopeDataByAttribute for array pattern
      *
      * @api
      * @param int $categoryId
